Implement inventory sales and device APIs

+++ backend/src/main/java/com/example/vendingmachine/model/Inventory.java
@@ -1,12 +1,14 @@
 package com.example.vendingmachine.model;
 
+import java.math.BigDecimal;
+
 public class Inventory {
 
     private Long inventoryId;
     private Long machineId;
     private Long drinkId;
     private Integer quantity;
-    private Integer price;
+    private BigDecimal price;
     private Integer lowStockThreshold;
     private Integer capacity;
 
@@ -18,10 +20,13 @@ public class Inventory {
     public void setDrinkId(Long drinkId) { this.drinkId = drinkId; }
     public Integer getQuantity() { return quantity; }
     public void setQuantity(Integer quantity) { this.quantity = quantity; }
-    public Integer getPrice() { return price; }
-    public void setPrice(Integer price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
+
+    // Java 欄位命名保留 lowStockThreshold，對應資料庫欄位 Inventory.threshold。
     public Integer getLowStockThreshold() { return lowStockThreshold; }
     public void setLowStockThreshold(Integer lowStockThreshold) { this.lowStockThreshold = lowStockThreshold; }
+
     public Integer getCapacity() { return capacity; }
     public void setCapacity(Integer capacity) { this.capacity = capacity; }
 }
